+++ test/k8s/client/runner/base_test.exs
@@ -15,10 +15,15 @@ defmodule K8s.Client.Runner.BaseTest do
 
     def request(:get, @namespaced_url, _, _, _), do: render(nil)
 
-    def request(:post, @namespaced_url, _, _, _), do: render(nil)
+    def request(:post, @namespaced_url, body, _, _), do: render(body)
 
     def request(:get, @namespaced_url <> "/test", _body, _headers, _opts) do
       render(nil)
+    end
+
+    def request(:get, @namespaced_url <> "/test-query-params", _body, _headers, opts) do
+      params = Keyword.get(opts, :params)
+      render(params)
     end
 
     def request(
@@ -45,80 +50,85 @@ defmodule K8s.Client.Runner.BaseTest do
     def request(
           :post,
           @base_url <> "/api/v1/namespaces/default/pods/nginx/eviction",
-          _body,
+          body,
           _headers,
           _opts
         ) do
-      render(nil)
+      render(body)
     end
   end
 
   setup do
     DynamicHTTPProvider.register(self(), __MODULE__.HTTPMock)
     conn = K8s.Conn.from_file("test/support/kube-config.yaml")
-    {:ok, cluster} = K8s.Cluster.Registry.add(:base_runner_test, conn)
 
-    {:ok, cluster: cluster}
+    {:ok, conn: conn}
   end
 
   describe "run/3" do
-    test "running an operation with a K8s.Selector set", %{cluster: cluster} do
+    test "running an operation with a K8s.Selector set", %{conn: conn} do
       operation =
         Client.list("v1", :pods)
         |> K8s.Selector.label({"app", "nginx"})
 
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
     end
 
-    test "running an operation without an HTTP body", %{cluster: cluster} do
+    test "running an operation without an HTTP body", %{conn: conn} do
       operation = Client.get(make_namespace("test"))
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
     end
 
-    test "running an operation with an HTTP body", %{cluster: cluster} do
+    test "running an operation with an HTTP body", %{conn: conn} do
       operation = Client.create(make_namespace("test"))
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
     end
 
-    test "running an operation with options", %{cluster: cluster} do
-      operation = Client.get(make_namespace("test"))
-      opts = [params: %{"watch" => "true"}]
-      assert {:ok, _} = Base.run(operation, cluster, opts)
+    test "running an operation with options", %{conn: conn} do
+      operation = Client.get(make_namespace("test-query-params"))
+      params = %{"watch" => "true"}
+      operation_w_params = Map.put(operation, :query_params, params)
+
+      assert {:ok, ^params} = Base.run(operation_w_params, conn)
     end
 
-    test "supports subresource operations", %{cluster: cluster} do
+    test "supports subresource operations", %{conn: conn} do
       operation = Client.get("apps/v1", "deployments/status", name: "nginx", namespace: "default")
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
+    end
+
+    test "running an operation with a custom HTTP body", %{conn: conn} do
+      operation = Client.create(make_namespace("test"))
+      labels = %{"env" => "test"}
+      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
+
+      assert {:ok, body} = Base.run(operation, conn, body)
+
+      assert body ==
+               ~s({"apiVersion":"v1","kind":"Namespace","metadata":{"labels":{"env":"test"},"name":"test"}})
     end
   end
 
   describe "run/4" do
-    test "running an operation with a custom HTTP body", %{cluster: cluster} do
-      operation = Client.create(make_namespace("test"))
-      labels = %{"env" => "test"}
-      body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
-
-      assert {:ok, _} = Base.run(operation, cluster, body)
-    end
-
-    test "running an operation with a custom HTTP body and options", %{
-      cluster: cluster
+    test "[DEPRECATED] running an operation with a custom HTTP body and options", %{
+      conn: conn
     } do
       operation = Client.create(make_namespace("test"))
       labels = %{"env" => "test"}
       body = put_in(make_namespace("test"), ["metadata", "labels"], labels)
+
       opts = [params: %{"watch" => "true"}]
-      assert {:ok, _} = Base.run(operation, cluster, body, opts)
+      assert {:ok, _} = Base.run(operation, conn, body, opts)
     end
   end
 
   describe "run" do
-    test "request with HTTP 2xx response", %{cluster: cluster} do
+    test "request with HTTP 2xx response", %{conn: conn} do
       operation = Client.list("v1", "Namespace", [])
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
     end
 
-    test "supports subresource operations with alternate `kind` HTTP bodies", %{cluster: cluster} do
+    test "supports subresource operations with alternate `kind` HTTP bodies", %{conn: conn} do
       pod = %{
         "apiVersion" => "v1",
         "kind" => "Pod",
@@ -138,7 +148,7 @@ defmodule K8s.Client.Runner.BaseTest do
       }
 
       operation = K8s.Client.create(pod, eviction)
-      assert {:ok, _} = Base.run(operation, cluster)
+      assert {:ok, _} = Base.run(operation, conn)
     end
   end
 end