+++ test/k8s/middleware/request/initialize_test.exs
@@ -1,0 +1,21 @@
+defmodule K8s.Middleware.Request.InitializeTest do
+  use ExUnit.Case, async: true
+
+  test "initializes a request headers from K8s.Conn.RequestOptions" do
+    conn = K8s.Conn.from_file("./test/support/kube-config.yaml")
+    request = %K8s.Middleware.Request{conn: conn}
+
+    {:ok, %{headers: headers}} = K8s.Middleware.Request.Initialize.call(request)
+
+    assert headers == [{"Accept", "application/json"}, {"Content-Type", "application/json"}]
+  end
+
+  test "initializes a HTTPoison options from K8s.Conn.RequestOptions" do
+    conn = K8s.Conn.from_file("./test/support/kube-config.yaml")
+    request = %K8s.Middleware.Request{conn: conn}
+
+    {:ok, %{opts: opts}} = K8s.Middleware.Request.Initialize.call(request)
+
+    assert Keyword.has_key?(opts, :ssl)
+  end
+end